requirements.txt
# Api
uvicorn==0.34.2
starlette==0.46.2
fastapi==0.115.12
python-multipart==0.0.20
jinja2==3.1.6
tqdm==4.67.1
filemagic==1.6
pytesseract==0.3.13
pillow==11.2.1
opencv-python==4.11.0.86
numpy==2.2.5

yadisk==3.3.0
httpx==0.28.1

aiofiles==24.1.0
aiohttp==3.11.18

# Config
pyyaml==6.0.2
pydantic==2.11.4

# Pipeline
ffmpeg-python==0.2.0
openai-whisper==20240930
gigachat==0.1.39.post1
